fix hmr (move appContext in its own file) & tailwind header
