+++ apps/frontend/src/components/ui/Card.tsx
@@ -4,7 +4,7 @@ export function Card({ children, className, ...props }: React.HTMLProps<HTMLDivE
   return (
     <div
       className={clsx(
-        "border border-neutral-300 bg-white p-4 shadow-sm dark:border-neutral-700 dark:bg-neutral-900",
+        "border border-neutral-300 bg-white p-4 shadow-sm dark:border-neutral-700 dark:bg-neutral-800",
         className,
         { ...props },
       )}
