+++ apps/frontend/src/components/views/RunnerDetailsView.tsx
@@ -6,6 +6,7 @@ import { Col, Row } from "react-bootstrap";
 import { Link, useNavigate, useParams } from "react-router-dom";
 import { stringUtils } from "@live24hisere/utils";
 import { SearchParam } from "../../constants/searchParams";
+import { appContext } from "../../contexts/AppContext";
 import { useGetPublicEditions } from "../../hooks/api/requests/public/editions/useGetPublicEditions";
 import { useGetPublicRunnerParticipations } from "../../hooks/api/requests/public/participants/useGetPublicRunnerParticipations";
 import { useGetPublicRaces } from "../../hooks/api/requests/public/races/useGetPublicRaces";
@@ -18,7 +19,6 @@ import { getCountryAlpha2CodeFromAlpha3Code } from "../../utils/countryUtils";
 import { generateXlsxFromData } from "../../utils/excelUtils";
 import { isRaceFinished } from "../../utils/raceUtils";
 import { getDataForExcelExport, getRaceRunnerFromRunnerAndParticipant } from "../../utils/runnerUtils";
-import { appContext } from "../App";
 import { Card } from "../ui/Card";
 import CircularLoader from "../ui/CircularLoader";
 import { Flag } from "../ui/countries/Flag";
