+++ apps/frontend/src/components/views/admin/runners/ImportRunnersCsvView.tsx
@@ -7,12 +7,12 @@ import type { Gender, RaceRunner } from "@live24hisere/core/types";
 import { stringUtils } from "@live24hisere/utils";
 import { GENDER_OPTIONS } from "../../../../constants/forms";
 import { ImportCsvColumn } from "../../../../constants/importCsv";
+import { appContext } from "../../../../contexts/AppContext";
 import { useGetAdminRaces } from "../../../../hooks/api/requests/admin/races/useGetAdminRaces";
 import { getImportRunnersCsvBreadcrumbs } from "../../../../services/breadcrumbs/breadcrumbService";
 import type { SelectOption } from "../../../../types/Forms";
 import type { RunnerFromCsv, RunnersCsvMapping } from "../../../../types/ImportCsv";
 import { getRunnerFromCsv, parseCsv } from "../../../../utils/csvUtils";
-import { appContext } from "../../../App";
 import { Card } from "../../../ui/Card";
 import { Input } from "../../../ui/forms/Input";
 import Select from "../../../ui/forms/Select";
