+++ apps/frontend/src/hooks/api/useAccessToken.ts
@@ -1,5 +1,5 @@
 import React from "react";
-import { appContext } from "../../components/App";
+import { appContext } from "../../contexts/AppContext";
 
 export function useAccessToken(): string | null {
   return React.useContext(appContext).user.accessToken;
