+++ apps/frontend/src/hooks/useRanking.ts
@@ -1,6 +1,6 @@
 import React from "react";
 import type { PublicRace } from "@live24hisere/core/types";
-import { appContext } from "../components/App";
+import { appContext } from "../contexts/AppContext";
 import { RankingCalculator } from "../services/RankingCalculator";
 import type { MinimalRankingRunnerInput, Ranking } from "../types/Ranking";
