+++ eslint.config.js
@@ -71,6 +71,7 @@ export default [
       "guard-for-in": "off", // TODO enable ?
       "max-lines": "off",
       "no-alert": "off",
+      "no-octal-escape": "off", // Only octal escape works with tailwind
       "promise/avoid-new": "off",
       "react/no-unescaped-entities": "off",
       "@typescript-eslint/ban-ts-comment": "off",
